Update tripId type as Long
Why: as Dao returns Long whenever a new record is inserted

+++ app/src/main/java/com/gowthamraj07/journeytracker/data/PlacesRepositoryImpl.kt
@@ -17,7 +17,7 @@ class PlacesRepositoryImpl(
     private val flickrResponseParser: FlickrResponseParser,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : PlacesRepository {
-    override suspend fun loadPlacesFor(tripId: Int): Flow<List<Place>> {
+    override suspend fun loadPlacesFor(tripId: Long): Flow<List<Place>> {
         return placeDao.getPlacesByTrip(tripId).flowOn(ioDispatcher).map { places ->
             places.map { place ->
                 val flickrJsonResponse =
